refactor(GameController.cs): move game logic to helper classes for better code organization feat(ScoreController.cs): add gameId parameter to AddScore method to calculate score based on game feat(TileController.cs): create new controller to handle tile-specific actions feat(GameHelper.cs, ScoreHelper.cs, TileHelper.cs): create helper classes to handle game, score, and tile logic fix(Score.cs): change HighScore type from int to double for more precise scoring

+++ Models/Score.cs
@@ -8,7 +8,7 @@ namespace SPAmineseweeper.Models
         [Key]
         public int Id { get; set; }
         [Required]
-        public int HighScore { get; set; }
+        public double HighScore { get; set; }
 
         [ForeignKey("User")]
         public string? UserId { get; set; }
